added types tests
